reorganized screens to match Manav's folder structure
